Add code for new read in
