+++ pymagicc/MAGICC6/run/EDGAR_NOXI_EMIS.IN
