+++ pymagicc/MAGICC6/run/HISTRCP_HFC125_CONC.IN
@@ -2,14 +2,14 @@
 __RCP Historical Concentration File: HFC125__
 
 Data: Global average mixing ratio
-Gas: HFC125	
+Gas: HFC125
 Unit: Parts per trillion (ppt)
-Description: Input with historical mixing ratio data, based on Oram et al. (1998) and figure 1-23 WMO 2006, linearly interpolated to zero between 1980 and 1970	
+Description: Input with historical mixing ratio data, based on Oram et al. (1998) and figure 1-23 WMO 2006, linearly interpolated to zero between 1980 and 1970
 Source: WMO 2007, Fig 1-23
 Compiled by: Guus Velders, Steve Smith, Malte Meinshausen
 Date: 03-Aug-2009 17:23:01
 
-&THISFILE_SPECIFICATIONS 
+&THISFILE_SPECIFICATIONS
  THISFILE_DATACOLUMNS    =    1  ,
  THISFILE_FIRSTYEAR      = 1930  ,
  THISFILE_LASTYEAR       = 2004  ,
@@ -17,81 +17,81 @@ Date: 03-Aug-2009 17:23:01
  THISFILE_FIRSTDATAROW   =   23  ,
  THISFILE_UNITS          = ppt  ,
  THISFILE_DATTYPE        = STANDARD  ,
-/  
+/
 
-    COLCODE               YEARS
-       1930     0.00000000e+000 
-       1931     0.00000000e+000 
-       1932     0.00000000e+000 
-       1933     0.00000000e+000 
-       1934     0.00000000e+000 
-       1935     0.00000000e+000 
-       1936     0.00000000e+000 
-       1937     0.00000000e+000 
-       1938     0.00000000e+000 
-       1939     0.00000000e+000 
-       1940     0.00000000e+000 
-       1941     0.00000000e+000 
-       1942     0.00000000e+000 
-       1943     0.00000000e+000 
-       1944     0.00000000e+000 
-       1945     0.00000000e+000 
-       1946     0.00000000e+000 
-       1947     0.00000000e+000 
-       1948     0.00000000e+000 
-       1949     0.00000000e+000 
-       1950     0.00000000e+000 
-       1951     0.00000000e+000 
-       1952     0.00000000e+000 
-       1953     0.00000000e+000 
-       1954     0.00000000e+000 
-       1955     0.00000000e+000 
-       1956     0.00000000e+000 
-       1957     0.00000000e+000 
-       1958     0.00000000e+000 
-       1959     0.00000000e+000 
-       1960     0.00000000e+000 
-       1961     0.00000000e+000 
-       1962     0.00000000e+000 
-       1963     0.00000000e+000 
-       1964     0.00000000e+000 
-       1965     0.00000000e+000 
-       1966     0.00000000e+000 
-       1967     0.00000000e+000 
-       1968     0.00000000e+000 
-       1969     0.00000000e+000 
-       1970     0.00000000e+000 
-       1971     2.92000000e-003 
-       1972     5.84000000e-003 
-       1973     8.76000000e-003 
-       1974     1.16800000e-002 
-       1975     1.46000000e-002 
-       1976     1.75200000e-002 
-       1977     2.04400000e-002 
-       1978     2.33600000e-002 
-       1979     2.62800000e-002 
-       1980     2.92000000e-002 
-       1981     2.92000000e-002 
-       1982     3.17000000e-002 
-       1983     3.45000000e-002 
-       1984     3.72000000e-002 
-       1985     3.69000000e-002 
-       1986     3.14000000e-002 
-       1987     3.43000000e-002 
-       1988     4.25000000e-002 
-       1989     5.06000000e-002 
-       1990     5.88000000e-002 
-       1991     6.66000000e-002 
-       1992     7.44000000e-002 
-       1993     8.71000000e-002 
-       1994     1.04000000e-001 
-       1995     1.96600000e-001 
-       1996     3.46900000e-001 
-       1997     5.15200000e-001 
-       1998     7.79700000e-001 
-       1999     1.13260000e+000 
-       2000     1.52350000e+000 
-       2001     1.92130000e+000 
-       2002     2.34480000e+000 
-       2003     2.86190000e+000 
-       2004     3.37900000e+000 
+    COLCODE              GLOBAL
+       1930     0.00000000e+000
+       1931     0.00000000e+000
+       1932     0.00000000e+000
+       1933     0.00000000e+000
+       1934     0.00000000e+000
+       1935     0.00000000e+000
+       1936     0.00000000e+000
+       1937     0.00000000e+000
+       1938     0.00000000e+000
+       1939     0.00000000e+000
+       1940     0.00000000e+000
+       1941     0.00000000e+000
+       1942     0.00000000e+000
+       1943     0.00000000e+000
+       1944     0.00000000e+000
+       1945     0.00000000e+000
+       1946     0.00000000e+000
+       1947     0.00000000e+000
+       1948     0.00000000e+000
+       1949     0.00000000e+000
+       1950     0.00000000e+000
+       1951     0.00000000e+000
+       1952     0.00000000e+000
+       1953     0.00000000e+000
+       1954     0.00000000e+000
+       1955     0.00000000e+000
+       1956     0.00000000e+000
+       1957     0.00000000e+000
+       1958     0.00000000e+000
+       1959     0.00000000e+000
+       1960     0.00000000e+000
+       1961     0.00000000e+000
+       1962     0.00000000e+000
+       1963     0.00000000e+000
+       1964     0.00000000e+000
+       1965     0.00000000e+000
+       1966     0.00000000e+000
+       1967     0.00000000e+000
+       1968     0.00000000e+000
+       1969     0.00000000e+000
+       1970     0.00000000e+000
+       1971     2.92000000e-003
+       1972     5.84000000e-003
+       1973     8.76000000e-003
+       1974     1.16800000e-002
+       1975     1.46000000e-002
+       1976     1.75200000e-002
+       1977     2.04400000e-002
+       1978     2.33600000e-002
+       1979     2.62800000e-002
+       1980     2.92000000e-002
+       1981     2.92000000e-002
+       1982     3.17000000e-002
+       1983     3.45000000e-002
+       1984     3.72000000e-002
+       1985     3.69000000e-002
+       1986     3.14000000e-002
+       1987     3.43000000e-002
+       1988     4.25000000e-002
+       1989     5.06000000e-002
+       1990     5.88000000e-002
+       1991     6.66000000e-002
+       1992     7.44000000e-002
+       1993     8.71000000e-002
+       1994     1.04000000e-001
+       1995     1.96600000e-001
+       1996     3.46900000e-001
+       1997     5.15200000e-001
+       1998     7.79700000e-001
+       1999     1.13260000e+000
+       2000     1.52350000e+000
+       2001     1.92130000e+000
+       2002     2.34480000e+000
+       2003     2.86190000e+000
+       2004     3.37900000e+000
