+++ pymagicc/MAGICC6/run/HISTRCP_HFC134a_CONC.IN
@@ -2,14 +2,14 @@
 __RCP Historical Concentration File: HFC134a__
 
 Data: Global average mixing ratio
-Gas: HFC134a	
-Unit: Parts per trillion (ppt) 	
-Description: Input with historic end-of-year data - note that the Montzka et al. NOAA/ESRL/GMD data only extended to July 2008. All given concentrations are end-of-year values, assumed equal with the following years start-of-year values. 	
+Gas: HFC134a
+Unit: Parts per trillion (ppt)
+Description: Input with historic end-of-year data - note that the Montzka et al. NOAA/ESRL/GMD data only extended to July 2008. All given concentrations are end-of-year values, assumed equal with the following years start-of-year values.
 Source: (NOAA/ESRL/GMD), available here ftp://ftp.cmdl.noaa.gov/hats/hfcs/134A_MS2008a.txt; Montzka et al. (1996), Observations of HFC-134a in the remote troposphere, GRL, 23, 169-172, 1996.  and Montzka et al. (1996) Decline in the tropospheric abundance of halogen from halocarbons: Implications for stratospheric ozone depletion. Sent by Steve Montzka in email on Thu, 07 May 2009 13:48:33 -0600
 Compiled by: Steve Montzka, Steve Smith, Malte Meinshausen, malte.meinshausen@pik-potsdam.de
 Date: 03-Aug-2009 09:37:46
 
-&THISFILE_SPECIFICATIONS 
+&THISFILE_SPECIFICATIONS
  THISFILE_DATACOLUMNS    =    1  ,
  THISFILE_FIRSTYEAR      = 1930  ,
  THISFILE_LASTYEAR       = 2007  ,
@@ -17,84 +17,84 @@ Date: 03-Aug-2009 09:37:46
  THISFILE_FIRSTDATAROW   =   23  ,
  THISFILE_UNITS          = ppt  ,
  THISFILE_DATTYPE        = STANDARD  ,
-/  
+/
 
-    COLCODE               YEARS
-       1930     0.00000000e+000 
-       1931     0.00000000e+000 
-       1932     0.00000000e+000 
-       1933     0.00000000e+000 
-       1934     0.00000000e+000 
-       1935     0.00000000e+000 
-       1936     0.00000000e+000 
-       1937     0.00000000e+000 
-       1938     0.00000000e+000 
-       1939     0.00000000e+000 
-       1940     0.00000000e+000 
+    COLCODE              GLOBAL
+       1930     0.00000000e+000
+       1931     0.00000000e+000
+       1932     0.00000000e+000
+       1933     0.00000000e+000
+       1934     0.00000000e+000
+       1935     0.00000000e+000
+       1936     0.00000000e+000
+       1937     0.00000000e+000
+       1938     0.00000000e+000
+       1939     0.00000000e+000
+       1940     0.00000000e+000
        1941     0.00000000e+000
-       1942     0.00000000e+000 
-       1943     0.00000000e+000 
-       1944     0.00000000e+000 
-       1945     0.00000000e+000 
-       1946     0.00000000e+000 
-       1947     0.00000000e+000 
-       1948     0.00000000e+000 
-       1949     0.00000000e+000 
-       1950     0.00000000e+000 
-       1951     0.00000000e+000 
-       1952     0.00000000e+000 
-       1953     0.00000000e+000 
-       1954     0.00000000e+000 
-       1955     0.00000000e+000 
-       1956     0.00000000e+000 
-       1957     0.00000000e+000 
-       1958     0.00000000e+000 
-       1959     0.00000000e+000 
-       1960     0.00000000e+000 
-       1961     0.00000000e+000 
-       1962     0.00000000e+000 
-       1963     0.00000000e+000 
-       1964     0.00000000e+000 
-       1965     0.00000000e+000 
-       1966     0.00000000e+000 
-       1967     0.00000000e+000 
-       1968     0.00000000e+000 
-       1969     0.00000000e+000 
-       1970     0.00000000e+000 
-       1971     0.00000000e+000 
-       1972     0.00000000e+000 
-       1973     0.00000000e+000 
-       1974     0.00000000e+000 
-       1975     0.00000000e+000 
-       1976     0.00000000e+000 
-       1977     0.00000000e+000 
-       1978     0.00000000e+000 
-       1979     0.00000000e+000 
-       1980     0.00000000e+000 
-       1981     0.00000000e+000 
-       1982     0.00000000e+000 
-       1983     0.00000000e+000 
-       1984     0.00000000e+000 
-       1985     0.00000000e+000 
-       1986     0.00000000e+000 
-       1987     0.00000000e+000 
-       1988     0.00000000e+000 
-       1989     0.00000000e+000 
-       1990     6.00000000e-002 
-       1991     1.20000000e-001 
-       1992     2.50000000e-001 
-       1993     5.00000000e-001 
-       1994     1.00000000e+000 
-       1995     2.29000000e+000 
-       1996     4.02000000e+000 
-       1997     6.32000000e+000 
-       1998     9.10000000e+000 
-       1999     1.22600000e+001 
-       2000     1.56700000e+001 
-       2001     1.93500000e+001 
-       2002     2.33800000e+001 
-       2003     2.75000000e+001 
-       2004     3.20500000e+001 
-       2005     3.66700000e+001 
-       2006     4.09400000e+001 
-       2007     4.53900000e+001 
+       1942     0.00000000e+000
+       1943     0.00000000e+000
+       1944     0.00000000e+000
+       1945     0.00000000e+000
+       1946     0.00000000e+000
+       1947     0.00000000e+000
+       1948     0.00000000e+000
+       1949     0.00000000e+000
+       1950     0.00000000e+000
+       1951     0.00000000e+000
+       1952     0.00000000e+000
+       1953     0.00000000e+000
+       1954     0.00000000e+000
+       1955     0.00000000e+000
+       1956     0.00000000e+000
+       1957     0.00000000e+000
+       1958     0.00000000e+000
+       1959     0.00000000e+000
+       1960     0.00000000e+000
+       1961     0.00000000e+000
+       1962     0.00000000e+000
+       1963     0.00000000e+000
+       1964     0.00000000e+000
+       1965     0.00000000e+000
+       1966     0.00000000e+000
+       1967     0.00000000e+000
+       1968     0.00000000e+000
+       1969     0.00000000e+000
+       1970     0.00000000e+000
+       1971     0.00000000e+000
+       1972     0.00000000e+000
+       1973     0.00000000e+000
+       1974     0.00000000e+000
+       1975     0.00000000e+000
+       1976     0.00000000e+000
+       1977     0.00000000e+000
+       1978     0.00000000e+000
+       1979     0.00000000e+000
+       1980     0.00000000e+000
+       1981     0.00000000e+000
+       1982     0.00000000e+000
+       1983     0.00000000e+000
+       1984     0.00000000e+000
+       1985     0.00000000e+000
+       1986     0.00000000e+000
+       1987     0.00000000e+000
+       1988     0.00000000e+000
+       1989     0.00000000e+000
+       1990     6.00000000e-002
+       1991     1.20000000e-001
+       1992     2.50000000e-001
+       1993     5.00000000e-001
+       1994     1.00000000e+000
+       1995     2.29000000e+000
+       1996     4.02000000e+000
+       1997     6.32000000e+000
+       1998     9.10000000e+000
+       1999     1.22600000e+001
+       2000     1.56700000e+001
+       2001     1.93500000e+001
+       2002     2.33800000e+001
+       2003     2.75000000e+001
+       2004     3.20500000e+001
+       2005     3.66700000e+001
+       2006     4.09400000e+001
+       2007     4.53900000e+001
