+++ pymagicc/MAGICC6/run/HISTRCP_HFC143a_CONC.IN
@@ -3,13 +3,13 @@ __RCP Historical Concentration File: HFC143a__
 
 Data: Global average mixing ratio, end of year
 Gas: HFC143a
-Unit: Parts per trillion (ppt) 
+Unit: Parts per trillion (ppt)
 Description: Input with historic data taken from Culbertson et al.,  Table 1 (interpolated end-of-year values between 1978 and 1996); linear interpolation to assumed zero level in 1970 and linear extrapolation to year 2000
-Source: Culbertson, et al., Chemosphere 55, 1109-1119, 2004 
+Source: Culbertson, et al., Chemosphere 55, 1109-1119, 2004
 Compiled by: Malte Meinshausen, malte.meinshausen@pik-potsdam.de, Steve Smith
 Date: 12-Aug-2009 00:14:09
 
-&THISFILE_SPECIFICATIONS 
+&THISFILE_SPECIFICATIONS
  THISFILE_DATACOLUMNS    =    1  ,
  THISFILE_FIRSTYEAR      = 1930  ,
  THISFILE_LASTYEAR       = 2000  ,
@@ -17,77 +17,77 @@ Date: 12-Aug-2009 00:14:09
  THISFILE_FIRSTDATAROW   =   23  ,
  THISFILE_UNITS          = ppt  ,
  THISFILE_DATTYPE        = STANDARD  ,
-/  
+/
 
-    COLCODE               YEARS
-       1930     0.00000000e+000 
-       1931     0.00000000e+000 
-       1932     0.00000000e+000 
-       1933     0.00000000e+000 
-       1934     0.00000000e+000 
-       1935     0.00000000e+000 
-       1936     0.00000000e+000 
-       1937     0.00000000e+000 
-       1938     0.00000000e+000 
-       1939     0.00000000e+000 
-       1940     0.00000000e+000 
-       1941     0.00000000e+000 
-       1942     0.00000000e+000 
-       1943     0.00000000e+000 
-       1944     0.00000000e+000 
-       1945     0.00000000e+000 
-       1946     0.00000000e+000 
-       1947     0.00000000e+000 
-       1948     0.00000000e+000 
-       1949     0.00000000e+000 
-       1950     0.00000000e+000 
-       1951     0.00000000e+000 
-       1952     0.00000000e+000 
-       1953     0.00000000e+000 
-       1954     0.00000000e+000 
-       1955     0.00000000e+000 
-       1956     0.00000000e+000 
-       1957     0.00000000e+000 
-       1958     0.00000000e+000 
-       1959     0.00000000e+000 
-       1960     0.00000000e+000 
-       1961     0.00000000e+000 
-       1962     0.00000000e+000 
-       1963     0.00000000e+000 
-       1964     0.00000000e+000 
-       1965     0.00000000e+000 
-       1966     0.00000000e+000 
-       1967     0.00000000e+000 
-       1968     0.00000000e+000 
-       1969     0.00000000e+000 
-       1970     0.00000000e+000 
-       1971     2.50000000e-002 
-       1972     5.00000000e-002 
-       1973     7.50000000e-002 
-       1974     1.00000000e-001 
-       1975     1.25000000e-001 
-       1976     1.50000000e-001 
-       1977     1.75000000e-001 
-       1978     2.00000000e-001 
-       1979     2.00000000e-001 
-       1980     2.00000000e-001 
-       1981     2.50000000e-001 
-       1982     3.00000000e-001 
-       1983     3.00000000e-001 
-       1984     3.50000000e-001 
-       1985     4.00000000e-001 
-       1986     4.00000000e-001 
-       1987     4.50000000e-001 
-       1988     5.50000000e-001 
-       1989     6.00000000e-001 
-       1990     6.50000000e-001 
-       1991     7.00000000e-001 
-       1992     7.50000000e-001 
-       1993     8.50000000e-001 
-       1994     1.00000000e+000 
-       1995     1.25000000e+000 
-       1996     1.65000000e+000 
-       1997     2.05000000e+000 
-       1998     2.45000000e+000 
-       1999     2.85000000e+000 
-       2000     3.25000000e+000 
+    COLCODE              GLOBAL
+       1930     0.00000000e+000
+       1931     0.00000000e+000
+       1932     0.00000000e+000
+       1933     0.00000000e+000
+       1934     0.00000000e+000
+       1935     0.00000000e+000
+       1936     0.00000000e+000
+       1937     0.00000000e+000
+       1938     0.00000000e+000
+       1939     0.00000000e+000
+       1940     0.00000000e+000
+       1941     0.00000000e+000
+       1942     0.00000000e+000
+       1943     0.00000000e+000
+       1944     0.00000000e+000
+       1945     0.00000000e+000
+       1946     0.00000000e+000
+       1947     0.00000000e+000
+       1948     0.00000000e+000
+       1949     0.00000000e+000
+       1950     0.00000000e+000
+       1951     0.00000000e+000
+       1952     0.00000000e+000
+       1953     0.00000000e+000
+       1954     0.00000000e+000
+       1955     0.00000000e+000
+       1956     0.00000000e+000
+       1957     0.00000000e+000
+       1958     0.00000000e+000
+       1959     0.00000000e+000
+       1960     0.00000000e+000
+       1961     0.00000000e+000
+       1962     0.00000000e+000
+       1963     0.00000000e+000
+       1964     0.00000000e+000
+       1965     0.00000000e+000
+       1966     0.00000000e+000
+       1967     0.00000000e+000
+       1968     0.00000000e+000
+       1969     0.00000000e+000
+       1970     0.00000000e+000
+       1971     2.50000000e-002
+       1972     5.00000000e-002
+       1973     7.50000000e-002
+       1974     1.00000000e-001
+       1975     1.25000000e-001
+       1976     1.50000000e-001
+       1977     1.75000000e-001
+       1978     2.00000000e-001
+       1979     2.00000000e-001
+       1980     2.00000000e-001
+       1981     2.50000000e-001
+       1982     3.00000000e-001
+       1983     3.00000000e-001
+       1984     3.50000000e-001
+       1985     4.00000000e-001
+       1986     4.00000000e-001
+       1987     4.50000000e-001
+       1988     5.50000000e-001
+       1989     6.00000000e-001
+       1990     6.50000000e-001
+       1991     7.00000000e-001
+       1992     7.50000000e-001
+       1993     8.50000000e-001
+       1994     1.00000000e+000
+       1995     1.25000000e+000
+       1996     1.65000000e+000
+       1997     2.05000000e+000
+       1998     2.45000000e+000
+       1999     2.85000000e+000
+       2000     3.25000000e+000
