+++ pymagicc/MAGICC6/run/HISTRCP_HFC245fa_CONC.IN
@@ -1,7 +1,7 @@
 
 __RCP Historical Concentration File: HFC245fa__
 
-Data: Global average mixing ratio 
+Data: Global average mixing ratio by näme with ümlauts ëh ça
 Gas: HFC245fa
 Unit: Parts per trillion (ppt)
 Description: Historical mixing ratio before 2000 assumed zero. HFC245fa, rather than HFC245ca, is the actually used isomer.
@@ -9,83 +9,83 @@ Source: See as well http://data.giss.nasa.gov/modelforce/ghgases/TG_A.1992-2008.
 Compiled by: Malte Meinshausen, malte.meinshausen@pik-potsdam.de, Steve Smith
 Date: 18-Jul-2006 10:44:17
 
-&THISFILE_SPECIFICATIONS 
+&THISFILE_SPECIFICATIONS
  THISFILE_DATACOLUMNS    = 1   ,
  THISFILE_FIRSTYEAR      = 1930  ,
  THISFILE_LASTYEAR       = 2000  ,
  THISFILE_ANNUALSTEPS    =    1  ,
  THISFILE_FIRSTDATAROW   =   21  ,
-/  
+/
 
        YEARS    GLOBAL_MIXINGRATIO
-       1930     0.00000000e+000  
-       1931     0.00000000e+000  
-       1932     0.00000000e+000  
-       1933     0.00000000e+000  
-       1934     0.00000000e+000  
-       1935     0.00000000e+000  
-       1936     0.00000000e+000  
-       1937     0.00000000e+000  
-       1938     0.00000000e+000  
-       1939     0.00000000e+000  
-       1940     0.00000000e+000  
-       1941     0.00000000e+000  
-       1942     0.00000000e+000  
-       1943     0.00000000e+000  
-       1944     0.00000000e+000  
-       1945     0.00000000e+000  
-       1946     0.00000000e+000  
-       1947     0.00000000e+000  
-       1948     0.00000000e+000  
-       1949     0.00000000e+000  
-       1950     0.00000000e+000  
-       1951     0.00000000e+000  
-       1952     0.00000000e+000  
-       1953     0.00000000e+000  
-       1954     0.00000000e+000  
-       1955     0.00000000e+000  
-       1956     0.00000000e+000  
-       1957     0.00000000e+000  
-       1958     0.00000000e+000  
-       1959     0.00000000e+000  
-       1960     0.00000000e+000  
-       1961     0.00000000e+000  
-       1962     0.00000000e+000  
-       1963     0.00000000e+000  
-       1964     0.00000000e+000  
-       1965     0.00000000e+000  
-       1966     0.00000000e+000  
-       1967     0.00000000e+000  
-       1968     0.00000000e+000  
-       1969     0.00000000e+000  
-       1970     0.00000000e+000  
-       1971     0.00000000e+000  
-       1972     0.00000000e+000  
-       1973     0.00000000e+000  
-       1974     0.00000000e+000  
-       1975     0.00000000e+000  
-       1976     0.00000000e+000  
-       1977     0.00000000e+000  
-       1978     0.00000000e+000  
-       1979     0.00000000e+000  
-       1980     0.00000000e+000  
-       1981     0.00000000e+000  
-       1982     0.00000000e+000  
-       1983     0.00000000e+000  
-       1984     0.00000000e+000  
-       1985     0.00000000e+000  
-       1986     0.00000000e+000  
-       1987     0.00000000e+000  
-       1988     0.00000000e+000  
-       1989     0.00000000e+000  
-       1990     0.00000000e+000  
-       1991     0.00000000e+000  
-       1992     0.00000000e+000  
-       1993     0.00000000e+000  
-       1994     0.00000000e+000  
-       1995     0.00000000e+000  
-       1996     0.00000000e+000  
-       1997     0.00000000e+000  
-       1998     0.00000000e+000  
-       1999     0.00000000e+000  
-       2000     0.00000000e+000  
+       1930     0.00000000e+000
+       1931     0.00000000e+000
+       1932     0.00000000e+000
+       1933     0.00000000e+000
+       1934     0.00000000e+000
+       1935     0.00000000e+000
+       1936     0.00000000e+000
+       1937     0.00000000e+000
+       1938     0.00000000e+000
+       1939     0.00000000e+000
+       1940     0.00000000e+000
+       1941     0.00000000e+000
+       1942     0.00000000e+000
+       1943     0.00000000e+000
+       1944     0.00000000e+000
+       1945     0.00000000e+000
+       1946     0.00000000e+000
+       1947     0.00000000e+000
+       1948     0.00000000e+000
+       1949     0.00000000e+000
+       1950     0.00000000e+000
+       1951     0.00000000e+000
+       1952     0.00000000e+000
+       1953     0.00000000e+000
+       1954     0.00000000e+000
+       1955     0.00000000e+000
+       1956     0.00000000e+000
+       1957     0.00000000e+000
+       1958     0.00000000e+000
+       1959     0.00000000e+000
+       1960     0.00000000e+000
+       1961     0.00000000e+000
+       1962     0.00000000e+000
+       1963     0.00000000e+000
+       1964     0.00000000e+000
+       1965     0.00000000e+000
+       1966     0.00000000e+000
+       1967     0.00000000e+000
+       1968     0.00000000e+000
+       1969     0.00000000e+000
+       1970     0.00000000e+000
+       1971     0.00000000e+000
+       1972     0.00000000e+000
+       1973     0.00000000e+000
+       1974     0.00000000e+000
+       1975     0.00000000e+000
+       1976     0.00000000e+000
+       1977     0.00000000e+000
+       1978     0.00000000e+000
+       1979     0.00000000e+000
+       1980     0.00000000e+000
+       1981     0.00000000e+000
+       1982     0.00000000e+000
+       1983     0.00000000e+000
+       1984     0.00000000e+000
+       1985     0.00000000e+000
+       1986     0.00000000e+000
+       1987     0.00000000e+000
+       1988     0.00000000e+000
+       1989     0.00000000e+000
+       1990     0.00000000e+000
+       1991     0.00000000e+000
+       1992     0.00000000e+000
+       1993     0.00000000e+000
+       1994     0.00000000e+000
+       1995     0.00000000e+000
+       1996     0.00000000e+000
+       1997     0.00000000e+000
+       1998     0.00000000e+000
+       1999     0.00000000e+000
+       2000     0.00000000e+000
