+++ pymagicc/MAGICC6/run/MAGTUNE_C4MIP_BERN.CFG
@@ -1,19 +1,19 @@
  !C4MIP_BERNcc
- &NML_ALLCFGS 
- CO2_DETRITUSPOOL_INITIAL =  9.277380e+001,  
-  CO2_FEEDBACKFACTOR_DETRITUS =  -1.357817e-001,  
-  CO2_FEEDBACKFACTOR_GPP =  1.070037e-002,  
-  CO2_FEEDBACKFACTOR_RESPIRATION =  6.845893e-002,  
-  CO2_FEEDBACKFACTOR_SOIL =  1.540879e-001,  
+ &NML_ALLCFGS
+ CO2_DETRITUSPOOL_INITIAL =  9.277380e+001,
+  CO2_FEEDBACKFACTOR_DETRITUS =  -1.357817e-001,
+  CO2_FEEDBACKFACTOR_GPP =  1.070037e-002,
+  CO2_FEEDBACKFACTOR_RESPIRATION =  6.845893e-002,
+  CO2_FEEDBACKFACTOR_SOIL =  1.540879e-001,
   CO2_FERTILIZATION_FACTOR =  6.485981e-001,
   CO2_FERTILIZATION_METHOD =  1.100486e+000,
   CO2_FRACTION_DETRITUS_2_SOIL =  1.007630e-003,
-  CO2_FRACTION_GPP_2_DETRITUS =  3.998165e-001,  
-  CO2_FRACTION_GPP_2_PLANT =  4.482615e-001,  
-  CO2_FRACTION_PLANT_2_DETRITUS =  9.989021e-001,  
-  CO2_GPP_INITIAL =  6.627160e+001,  
-  CO2_PLANTPOOL_INITIAL =  8.848584e+002,  
-  CO2_RESPIRATION_INITIAL =  1.226025e+001,  
+  CO2_FRACTION_GPP_2_DETRITUS =  3.998165e-001,
+  CO2_FRACTION_GPP_2_PLANT =  4.482615e-001,
+  CO2_FRACTION_PLANT_2_DETRITUS =  9.989021e-001,
+  CO2_GPP_INITIAL =  6.627160e+001,
+  CO2_PLANTPOOL_INITIAL =  8.848584e+002,
+  CO2_RESPIRATION_INITIAL =  1.226025e+001,
   CO2_SOILPOOL_INITIAL =  1.681525e+003,
   OCEANCC_SCALE_GASXCHANGE =  1.833492e+000,
   OCEANCC_SCALE_IMPULSERESPONSE =  9.492864e-001,
