+++ pymagicc/MAGICC6/run/MAGTUNE_FULLTUNE_DEFAULT.CFG
@@ -1,14 +1,14 @@
- !Medium CMIP3 Emulation with set 3K climate sensitivity 
- &NML_ALLCFGS 
- CORE_AMPLIFY_OCN2LAND_HEATXCHNG =  1.124688e+000,  
-  CORE_CLIMATESENSITIVITY =  3,  
-  CORE_FEEDBACK_QSENSITIVITY =  7.842035e-009,  
-  CORE_HEATXCHANGE_LANDOCEAN =  1.442816e+000,  
-  CORE_HEATXCHANGE_NORTHSOUTH =  3.115475e-001,  
-  CORE_HEMISFRACTION_NH_LAND =  4.969140e-001,  
-  CORE_HEMISFRACTION_SH_LAND =  2.486170e-001,  
-  CORE_RLO =  1.317229e+000,  
-  CORE_VERTICALDIFFUSIVITY =  1.103595e+000,  
-  CORE_VERTICALDIFF_TOP_DKDT =  -1.910108e-001,  
-  SLR_EXPANSION_SCALING =  1,  
- / 
+ !Medium CMIP3 Emulation with set 3K climate sensitivity
+ &NML_ALLCFGS
+ CORE_AMPLIFY_OCN2LAND_HEATXCHNG =  1.124688e+000,
+  CORE_CLIMATESENSITIVITY =  3,
+  CORE_FEEDBACK_QSENSITIVITY =  7.842035e-009,
+  CORE_HEATXCHANGE_LANDOCEAN =  1.442816e+000,
+  CORE_HEATXCHANGE_NORTHSOUTH =  3.115475e-001,
+  CORE_HEMISFRACTION_NH_LAND =  4.969140e-001,
+  CORE_HEMISFRACTION_SH_LAND =  2.486170e-001,
+  CORE_RLO =  1.317229e+000,
+  CORE_VERTICALDIFFUSIVITY =  1.103595e+000,
+  CORE_VERTICALDIFF_TOP_DKDT =  -1.910108e-001,
+  SLR_EXPANSION_SCALING =  1,
+ /
