+++ pymagicc/MAGICC6/run/MAGTUNE_FULLTUNE_MEDIUM_CMIP3_ECS3.CFG
@@ -1,4 +1,4 @@
-!Medium CMIP3 Emulation with set 3K climate sensitivity 
+!Medium CMIP3 Emulation with set 3K climate sensitivity
 &NML_ALLCFGS
   CORE_AMPLIFY_OCN2LAND_HEATXCHNG =  1.124688e+000,
   CORE_CLIMATESENSITIVITY =  3,
