+++ pymagicc/MAGICC6/run/MARLAND_CO2I_EMIS.IN
