+++ pymagicc/MAGICC6/run/RCPODS_WMO2006_MixingRatios_A1.prn
@@ -4,7 +4,7 @@ __MAGICC 6.X DATA FILE__
 
 Data: Global average mixing ratios
 Gas: ODS
-Unit: metric tons
+Unit: ppt
 Description: 1951-2100 Baseline mixing ratio file generated by John Daniel and Guus Velders for WMO2006, Chapter 8. (2/3/06); CH3CL updated to reflect MAGICC6 timeseries after 1955 and lower 2000 concentrations closer to 535ppt in line with
 Source: John Daniel, pers. Communication, 25th May 2009, WMO2006, Chapter 8; Butler et al. 1999 Fig6 for CCl before 1950; CH3Cl - calculations based on emissions with MAGICC63 and 1.3 year lifetime.
 Contact: M. Meinshausen, malte.meinshausen@pik-potsdam.de
