+++ pymagicc/default_config.nml
@@ -332,8 +332,8 @@
     file_n2on_emis = ''
     file_noxi_emis = 'HISTRCP_NOXI_EMIS.IN'
     file_noxb_emis = 'HISTRCP_NOXB_EMIS.IN'
-    file_nh3i_emis = 'EDGAR_NOX_EMIS_FOSSIL&IND.IN'
-    file_nh3b_emis = 'EDGAR_NOX_EMIS_LANDUSE.IN'
+    file_nh3i_emis = 'EDGAR_NOXI_EMIS.IN'
+    file_nh3b_emis = 'EDGAR_NOXB_EMIS.IN'
     file_nmvoci_emis = 'HISTRCP_NMVOCI_EMIS.IN'
     file_nmvocb_emis = 'HISTRCP_NMVOCB_EMIS.IN'
     file_soxi_emis = 'HISTRCP_SOXI_EMIS.IN'
