+++ pymagicc/definitions/magicc_dattype_regionmode_regions.csv
@@ -1,6 +1,6 @@
 Notes,THISFILE_DATTYPE,THISFILE_REGIONMODE,Region1,Region2,Region3,Region4,Region5,Region6,Region7,Region8,Region9,Region10,Region11,Region12,Region13,Region14,Region15,Region16,Region17,Region18,Region19,Region20
-DATTYPE not used by MAGICC.F90 for SCEN7. Order matters for non SCEN7 files.,REGIONDATA,SRES,OECD,REF,ASIA,ALM,,,,,,,,,,,,,,,,
-DATTYPE not used by MAGICC.F90 for SCEN7. Order matters for non SCEN7 files.,REGIONDATA,SRES,WORLD,OECD,REF,ASIA,ALM,,,,,,,,,,,,,,,
+DATTYPE not used by MAGICC.F90 for SCEN7. Order matters for non SCEN7 files.,REGIONDATA,SRES,OECD90,REF,ASIA,ALM,,,,,,,,,,,,,,,,
+DATTYPE not used by MAGICC.F90 for SCEN7. Order matters for non SCEN7 files.,REGIONDATA,SRES,WORLD,OECD90,REF,ASIA,ALM,,,,,,,,,,,,,,,
 DATTYPE not used by MAGICC.F90 for SCEN7. Order matters for non SCEN7 files.,REGIONDATA,RCP,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,,,,,,,,,,,,,,,
 DATTYPE not used by MAGICC.F90 for SCEN7. Order matters for non SCEN7 files.,REGIONDATA,RCP,WORLD,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,,,,,,,,,,,,,,
 DATTYPE not used by MAGICC.F90 for SCEN7. Order matters for non SCEN7 files.,REGIONDATA,RCPPLUSBUNKERS,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,BUNKERS,,,,,,,,,,,,,,
@@ -12,10 +12,18 @@ DATTYPE not used by MAGICC.F90 for SCEN7. Order matters for non SCEN7 files.,REG
 SCEN7 files only. Order doesn't matter and any other regions in the file are ignored. If you give this config to a non-SCEN7 file then MAGICC will assume it's global data based off the fact that it's only a single column and THISFILE_DATTYPE ~= 'REGIONDATA',SCEN7,WORLD,WORLD,,,,,,,,,,,,,,,,,,,
 SCEN7 files only. Order doesn't matter and any other regions in the file are ignored. If you give this config to a non-SCEN7 file then MAGICC will assume it's global data based off the fact that it's only a single column and THISFILE_DATTYPE ~= 'REGIONDATA',SCEN7,WORLD,GLOBAL,,,,,,,,,,,,,,,,,,,
 SCEN7 files only. Order doesn't matter and any other regions in the file are ignored. If you give this config to a non-SCEN7 file then MAGICC will assume it's global data based off the fact that it's only a single column and THISFILE_DATTYPE ~= 'REGIONDATA',SCEN7,WORLD,GL,,,,,,,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 5 columns and do nothing.,SCEN7,SRES,OECD90,REF,ASIA,ALM,,,,,,,,,,,,,,,,
 SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 5 columns and do nothing.,SCEN7,SRES,WORLD,OECD90,REF,ASIA,ALM,,,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 6 columns and do nothing.,SCEN7,RCP,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,,,,,,,,,,,,,,,
 SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 6 columns and do nothing.,SCEN7,RCP,WORLD,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 6 columns and do nothing.,SCEN7,RCPPLUSBUNKERS,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,BUNKERS,,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 6 columns and do nothing.,SCEN7,RCPPLUSBUNKERS,WORLD,R5OECD,R5REF,R5ASIA,R5MAF,R5LAM,BUNKERS,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSBUNKERS,R6OECD90,R6REF,R6ASIA,R6MAF,R6LAM,BUNKERS,,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSBUNKERS,WORLD,R6OECD90,R6REF,R6ASIA,R6MAF,R6LAM,BUNKERS,,,,,,,,,,,,,
+SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSSPLITBUNKERS,R6OECD90,R6REF,R6ASIA,R6MAF,R6LAM,AVIATION,SHIPPING,,,,,,,,,,,,,
 SCEN7 files. Order doesn't matter. WORLD values ignored. If you give this to a non-SCEN7 file then it will incorrectly guess you have supplied FOURBOX data then find 8 columns and do nothing.,SCEN7,RCPPLUSSPLITBUNKERS,WORLD,R6OECD90,R6REF,R6ASIA,R6MAF,R6LAM,AVIATION,SHIPPING,,,,,,,,,,,,
 Not SCEN7 or SCEN files. Order matters.,NOTUSED,FOURBOX,GLOBAL,,,,,,,,,,,,,,,,,,,
 Not SCEN7 or SCEN files. Order matters.,NOTUSED,FOURBOX,NHLAND,SHLAND,,,,,,,,,,,,,,,,,,
 Not SCEN7 or SCEN files. Order matters.,NOTUSED,FOURBOX,GLOBAL,NHLAND,SHLAND,,,,,,,,,,,,,,,,,
 Not SCEN7 or SCEN files. Order matters.,NOTUSED,FOURBOX,NHOCEAN,NHLAND,SHOCEAN,SHLAND,,,,,,,,,,,,,,,,
+Not SCEN7 or SCEN files. Order matters.,NOTUSED,HEMISPHERIC,NH,SH,,,,,,,,,,,,,,,,,,
